Added Old Neon
This is the first verison of The Neon programming language i ever wrote. It's written in JavaScript...
Also it's a single file (module) that exports the neon() function. This functions accepts 2 (+ 1 optional) parameters. code: string (the neon code), main: string (the name of the main class), debug?: bool (If true, prints debug info)